update cu ultima lectie
