--- src/main/java/ro/radu/curs/JavaApplication.java
@@ -1,12 +1,0 @@
-package ro.radu.curs;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class JavaApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(JavaApplication.class, args);
-	}
-}
