--- src/main/java/ro/radu/curs/dto/HelloRequest.java
@@ -1,12 +1,0 @@
-package ro.radu.curs.dto;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-import lombok.Data;
-
-@Data
-public class HelloRequest {
-
-    @JsonProperty("input")
-    String input;
-}
