--- src/main/java/ro/radu/curs/dto/PriceDto.java
@@ -1,11 +1,0 @@
-package ro.radu.curs.dto;
-
-import java.math.BigDecimal;
-
-import lombok.Data;
-
-@Data
-public class PriceDto {
-    private BigDecimal price;
-    private String currency;
-}
