--- src/main/java/ro/radu/curs/dto/SedanDto.java
@@ -1,8 +1,0 @@
-package ro.radu.curs.dto;
-
-import lombok.Data;
-
-@Data
-public class SedanDto extends CarDto {
-    private Integer doors;
-}
