--- src/main/java/ro/radu/curs/dto/SuvDto.java
@@ -1,8 +1,0 @@
-package ro.radu.curs.dto;
-
-import lombok.Data;
-
-@Data
-public class SuvDto extends CarDto {
-    private Integer transmission;
-}
