--- src/main/java/ro/radu/curs/dto/VanDto.java
@@ -1,8 +1,0 @@
-package ro.radu.curs.dto;
-
-import lombok.Data;
-
-@Data
-public class VanDto extends CarDto {
-    private String capacity;
-}
