--- src/main/java/ro/radu/curs/repository/CarRepository.java
@@ -1,15 +1,0 @@
-package ro.radu.curs.repository;
-
-import java.util.List;
-
-import org.springframework.stereotype.Repository;
-
-import ro.radu.curs.model.Car;
-
-@Repository
-public class CarRepository {
-
-    public List<Car> getAllCars() {
-        return MemoryRepository.carList;
-    }
-}
