--- src/main/java/ro/radu/curs/repository/MemoryRepository.java
@@ -1,12 +1,0 @@
-package ro.radu.curs.repository;
-
-import java.util.ArrayList;
-import java.util.List;
-
-import lombok.experimental.UtilityClass;
-import ro.radu.curs.model.Car;
-
-@UtilityClass
-public class MemoryRepository {
-    public static List<Car> carList = new ArrayList<>();
-}
