+++ src/main/java/ro/raffa/curs/JavaApplication.java
@@ -1,0 +1,12 @@
+package ro.raffa.curs;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class JavaApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(JavaApplication.class, args);
+	}
+}
