+++ src/main/java/ro/raffa/curs/configuration/SecurityConfiguration.java
@@ -1,0 +1,20 @@
+package ro.raffa.curs.configuration;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.web.SecurityFilterChain;
+
+@Configuration
+public class SecurityConfiguration {
+    @Bean
+    public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
+        http
+                .csrf().disable()
+                .authorizeHttpRequests((authorize) -> authorize
+                        .mvcMatchers("/**").permitAll()
+                        .anyRequest().permitAll());
+
+        return http.build();
+    }
+}
