+++ src/main/java/ro/raffa/curs/dto/CarDto.java
@@ -1,0 +1,12 @@
+package ro.raffa.curs.dto;
+
+import lombok.Data;
+
+@Data
+public class CarDto {
+    private String maker;
+    private String model;
+    private String color;
+    private Integer year;
+    private PriceDto price;
+}
