+++ src/main/java/ro/raffa/curs/dto/HelloRequest.java
@@ -1,0 +1,12 @@
+package ro.raffa.curs.dto;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+import lombok.Data;
+
+@Data
+public class HelloRequest {
+
+    @JsonProperty("input")
+    String input;
+}
