+++ src/main/java/ro/raffa/curs/dto/PriceDto.java
@@ -1,0 +1,11 @@
+package ro.raffa.curs.dto;
+
+import java.math.BigDecimal;
+
+import lombok.Data;
+
+@Data
+public class PriceDto {
+    private BigDecimal price;
+    private String currency;
+}
