+++ src/main/java/ro/raffa/curs/dto/SedanDto.java
@@ -1,0 +1,8 @@
+package ro.raffa.curs.dto;
+
+import lombok.Data;
+
+@Data
+public class SedanDto extends CarDto {
+    private Integer doors;
+}
