+++ src/main/java/ro/raffa/curs/dto/SuvDto.java
@@ -1,0 +1,8 @@
+package ro.raffa.curs.dto;
+
+import lombok.Data;
+
+@Data
+public class SuvDto extends CarDto {
+    private Integer transmission;
+}
