+++ src/main/java/ro/raffa/curs/dto/VanDto.java
@@ -1,0 +1,9 @@
+package ro.raffa.curs.dto;
+
+import lombok.Data;
+
+@Data
+public class VanDto extends CarDto {
+    private String capacity;
+}
+
