+++ src/main/java/ro/raffa/curs/repository/CarRepository.java
@@ -1,0 +1,15 @@
+package ro.raffa.curs.repository;
+
+import java.util.List;
+
+import org.springframework.stereotype.Repository;
+
+import ro.raffa.curs.model.Car;
+
+@Repository
+public class CarRepository {
+
+    public List<Car> getAllCars() {
+        return MemoryRepository.carList;
+    }
+}
