+++ src/main/java/ro/raffa/curs/repository/MemoryRepository.java
@@ -1,0 +1,12 @@
+package ro.raffa.curs.repository;
+
+import java.util.ArrayList;
+import java.util.List;
+
+import lombok.experimental.UtilityClass;
+import ro.raffa.curs.model.Car;
+
+@UtilityClass
+public class MemoryRepository {
+    public static List<Car> carList = new ArrayList<>();
+}
