--- src/test/java/ro/radu/curs/JavaApplicationTests.java
@@ -1,13 +1,0 @@
-package ro.radu.curs;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class JavaApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
