+++ src/test/java/ro/raffa/curs/JavaApplicationTests.java
@@ -1,0 +1,13 @@
+package ro.raffa.curs;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class JavaApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
